move target specific scripts to tools/targets/
